% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_processing.R
\name{comma_to_numeric}
\alias{comma_to_numeric}
\title{Convert numbers with commas in them to numeric}
\usage{
comma_to_numeric(x)
}
\arguments{
\item{x}{character or factor}
}
\value{
numeric representation of input
}
\description{
Convert numbers with commas in them to numeric
}
\examples{
nums <- c("123,456,789", "313,214.03")
comma_to_numeric(nums)
}
