% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{inf_to_na}
\alias{inf_to_na}
\alias{inf_to_na.default}
\alias{inf_to_na.list}
\title{Convert infinite values to NA}
\usage{
inf_to_na(x, ...)

\method{inf_to_na}{default}(x)

\method{inf_to_na}{list}(x)
}
\arguments{
\item{x}{data_frame, matrix, etc}
}
\value{
object of original type but with infinite values as NA
}
\description{
Convert infinite values to NA
}
\examples{
df <- dplyr::data_frame(
x = c(1, 4, NA),
y = c(Inf, 3,  5)
)
inf_to_na(df)
}
