% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_processing.R
\name{sentence_case}
\alias{sentence_case}
\title{Convert text to sentence case}
\usage{
sentence_case(x)
}
\arguments{
\item{x}{phrase/word that should only have 1 uppercase}
}
\value{
sentence case output
}
\description{
Convert text to sentence case
}
\examples{
sentence_case("SDklfJSKLDF")
sentence_case("OMG")
sentence_case("This is OKAY")
}
