% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_apply.R
\name{slide_apply}
\alias{slide_apply}
\title{Use apply over a sliding window of data}
\usage{
slide_apply(data, window, step = 1, fun)
}
\arguments{
\item{data}{Array of data to apply the function over.}

\item{window}{Numeric. Number of data points in each window.}

\item{step}{Numeric. Number of indices to step through data by.}

\item{fun}{Function to apply over each window.}
}
\value{
Array of the results from the function.
}
\description{
Applies a function to subsets of a given dataset.
Note: sets values without sufficient data to NA.
Using 'step' is particularly useful for leave-one-out analysis.
}
\examples{
toy <- runif(150, min=0, max=20)
slide_apply(toy, 30, 1, sd) # calculates 30 day moving standard deviation
}
\author{
\href{https://github.com/stillmatic}{Chris Hua}
}
